# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for ليلتك wedding venue platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /merchants/register:
    post:
      operationId: registerMerchant
      tags: [merchants]
      summary: Register as a merchant/venue owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchantInput"
      responses:
        "201":
          description: Merchant registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
  /merchants/{id}:
    get:
      operationId: getMerchant
      tags: [merchants]
      summary: Get merchant profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Merchant profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
    patch:
      operationId: updateMerchant
      tags: [merchants]
      summary: Update merchant profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchantUpdate"
      responses:
        "200":
          description: Updated merchant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
  /merchants/{id}/venues:
    get:
      operationId: listMerchantVenues
      tags: [merchants]
      summary: List all venues owned by a merchant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Merchant's venues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venue"
    post:
      operationId: createMerchantVenue
      tags: [merchants]
      summary: Add a new venue (merchant)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenueInput"
      responses:
        "201":
          description: Venue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
  /merchants/{id}/venues/{venueId}:
    patch:
      operationId: updateMerchantVenue
      tags: [merchants]
      summary: Update a venue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenueUpdate"
      responses:
        "200":
          description: Updated venue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
    delete:
      operationId: deleteMerchantVenue
      tags: [merchants]
      summary: Delete/deactivate a venue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /merchants/{id}/dashboard:
    get:
      operationId: getMerchantDashboard
      tags: [merchants]
      summary: Get merchant dashboard stats (wow endpoint)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantDashboard"
  /merchants/{id}/bookings:
    get:
      operationId: listMerchantBookings
      tags: [merchants]
      summary: List bookings for merchant's venues
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Merchant bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /merchants/{id}/bookings/{bookingId}:
    patch:
      operationId: updateMerchantBooking
      tags: [merchants]
      summary: Approve or reject a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchantBookingUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # ─── STORAGE ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Venue:
      type: object
      required: [id, nameAr, categorySlug, pricePerNight, capacity, district, rating, reviewCount, images, isFeatured, isAvailable]
      properties:
        id:
          type: integer
        nameAr:
          type: string
        categorySlug:
          type: string
        pricePerNight:
          type: number
        capacity:
          type: integer
        district:
          type: string
          description: Riyadh district in Arabic
        rating:
          type: number
        reviewCount:
          type: integer
        images:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        isAvailable:
          type: boolean
        amenities:
          type: array
          items:
            type: string
        services:
          type: array
          items:
            type: string
        description:
          type: ["string", "null"]
        merchantId:
          type: ["integer", "null"]
...
    VenueInput:
      type: object
      required: [nameAr, categorySlug, pricePerNight, capacity, district]
      properties:
        nameAr:
          type: string
        categorySlug:
          type: string
        pricePerNight:
          type: number
        capacity:
          type: integer
        district:
          type: string
        amenities:
          type: array
          items:
            type: string
        services:
          type: array
          items:
            type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
        latitude:
          type: number
        longitude:
          type: number
        address:
          type: string
        phone:
          type: string
    VenueUpdate:
      type: object
      properties:
        nameAr:
          type: string
        categorySlug:
          type: string
        pricePerNight:
          type: number
        capacity:
          type: integer
        district:
          type: string
        amenities:
          type: array
          items:
            type: string
        services:
          type: array
          items:
            type: string
        description:
          type: string
        isAvailable:
          type: boolean
        images:
          type: array
          items:
            type: string
    Booking:
      type: object
      required: [id, venueId, venueName, eventDate, guestCount, status, totalPrice, reviewToken, createdAt]
      properties:
        id:
          type: integer
        venueId:
          type: integer
        venueName:
          type: string
        venueImage:
          type: ["string", "null"]
        eventDate:
          type: string
        guestCount:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        totalPrice:
          type: number
        reviewToken:
          type: string
          description: Unguessable UUID token used to submit a post-stay review (one-time use)
        notes:
          type: ["string", "null"]
        groomName:
          type: ["string", "null"]
        brideName:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Merchant:
      type: object
      required: [id, ownerName, businessName, phone, email, crNumber, status, createdAt]
      properties:
        id:
          type: integer
        ownerName:
          type: string
        businessName:
          type: string
        phone:
          type: string
        email:
          type: string
        crNumber:
          type: string
          description: Commercial registration number (سجل تجاري)
        status:
          type: string
          enum: [pending, approved, suspended]
        logoUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    MerchantInput:
      type: object
      required: [ownerName, businessName, phone, email, crNumber, password]
      properties:
        ownerName:
          type: string
          minLength: 1
          maxLength: 100
        businessName:
          type: string
          minLength: 1
          maxLength: 200
        phone:
          type: string
          minLength: 1
          maxLength: 20
        email:
          type: string
          minLength: 1
          maxLength: 200
        crNumber:
          type: string
          minLength: 1
          maxLength: 50
        password:
          type: string
          minLength: 8
          maxLength: 128
        description:
          type: string
          maxLength: 2000
        logoUrl:
          type: string
          maxLength: 500
    MerchantUpdate:
      type: object
      properties:
        ownerName:
          type: string
          minLength: 1
          maxLength: 100
        businessName:
          type: string
          minLength: 1
          maxLength: 200
        phone:
          type: string
          minLength: 1
          maxLength: 20
        description:
          type: string
          maxLength: 2000
        logoUrl:
          type: string
          maxLength: 500
    MerchantDashboard:
      type: object
      required: [merchantId, totalVenues, totalBookings, totalRevenue, pendingBookings, confirmedBookings, thisMonthRevenue, avgRating]
      properties:
        merchantId:
          type: integer
        totalVenues:
          type: integer
        totalBookings:
          type: integer
        totalRevenue:
          type: number
        pendingBookings:
          type: integer
        confirmedBookings:
          type: integer
        thisMonthRevenue:
          type: number
        avgRating:
          type: number
        recentBookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
    MerchantBookingUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [confirmed, cancelled]
        notes:
          type: string
...
  securitySchemes:
    ManageBearerAuth:
      type: http
      scheme: bearer
      description: >
        Invitation management bearer token (manageToken).
        Send as `Authorization: Bearer <token>`. Never place this token in
        a URL path or query string — it must only travel as a request header.