# lib/api-zod/src/generated/types/venue.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for ليلتك wedding venue platform
 * OpenAPI spec version: 0.1.0
 */

export interface Venue {
  id: number;
  nameAr: string;
  categorySlug: string;
  pricePerNight: number;
  capacity: number;
  /** Riyadh district in Arabic */
  district: string;
  rating: number;
  reviewCount: number;
  images: string[];
  isFeatured: boolean;
  isAvailable: boolean;
  amenities?: string[];
  services?: string[];
  /** @nullable */
  description?: string | null;
  /** @nullable */
  merchantId?: number | null;
}