# lib/api-zod/src/generated/types/venueUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for ليلتك wedding venue platform
 * OpenAPI spec version: 0.1.0
 */

export interface VenueUpdate {
  nameAr?: string;
  categorySlug?: string;
  pricePerNight?: number;
  capacity?: number;
  district?: string;
  amenities?: string[];
  services?: string[];
  description?: string;
  isAvailable?: boolean;
  images?: string[];
}